+++ Phase0/src/equation_builders/AddBedmasEquationBuilder.java
@@ -1,14 +1,13 @@
-package worksheet_maker;
+package equation_builders;
 
-public class SubBedmasEquationBuilder extends BedmasEquationBuilder {
+public class AddBedmasEquationBuilder extends BedmasEquationBuilder {
     @Override
     public void buildOperator() {
-        bedmasEquation.setOperator("-");
+        bedmasEquation.setOperator("+");
     }
 
     /**
-     * Creates operands (first and second) in the subtraction bedmas equation's question. If not negAllowed, prevent
-     * operand 2 from being greater than operand 1.
+     * Creates operands (first and second) in the addition bedmas equation's question. No limitations.
      *
      * @param operandRange1 the absolute range of values that the first operand can be.
      * @param operandRange2 the absolute range of values that the second operand can be.
@@ -16,14 +15,11 @@ public class SubBedmasEquationBuilder extends BedmasEquationBuilder {
      */
     @Override
     public void buildOperands(int[] operandRange1, int[] operandRange2, boolean negAllowed) {
-        //TODO: Fix bad inputs (ex. operand2 range is greater than operand1)
         int operand1 = randomize(operandRange1);
-        int operand2;
+        int operand2 = randomize(operandRange2);
         if (negAllowed) {
             operand1 = makeNegativeRandom(operand1);
-            operand2 = makeNegativeRandom(randomize(operandRange2));
-        } else {
-            operand2 = randomize(operandRange2[0], operand1);
+            operand2 = makeNegativeRandom(operand2);
         }
         bedmasEquation.setOperand1(operand1);
         bedmasEquation.setOperand2(operand2);
