+++ Phase0/src/equation_builders/SubBedmasEquationBuilder.java
@@ -1,13 +1,14 @@
-package worksheet_maker;
+package equation_builders;
 
-public class AddBedmasEquationBuilder extends BedmasEquationBuilder {
+public class SubBedmasEquationBuilder extends BedmasEquationBuilder {
     @Override
     public void buildOperator() {
-        bedmasEquation.setOperator("+");
+        bedmasEquation.setOperator("-");
     }
 
     /**
-     * Creates operands (first and second) in the addition bedmas equation's question. No limitations.
+     * Creates operands (first and second) in the subtraction bedmas equation's question. If not negAllowed, prevent
+     * operand 2 from being greater than operand 1.
      *
      * @param operandRange1 the absolute range of values that the first operand can be.
      * @param operandRange2 the absolute range of values that the second operand can be.
@@ -15,11 +16,14 @@ public class AddBedmasEquationBuilder extends BedmasEquationBuilder {
      */
     @Override
     public void buildOperands(int[] operandRange1, int[] operandRange2, boolean negAllowed) {
+        //TODO: Fix bad inputs (ex. operand2 range is greater than operand1)
         int operand1 = randomize(operandRange1);
-        int operand2 = randomize(operandRange2);
+        int operand2;
         if (negAllowed) {
             operand1 = makeNegativeRandom(operand1);
-            operand2 = makeNegativeRandom(operand2);
+            operand2 = makeNegativeRandom(randomize(operandRange2));
+        } else {
+            operand2 = randomize(operandRange2[0], operand1);
         }
         bedmasEquation.setOperand1(operand1);
         bedmasEquation.setOperand2(operand2);
