+++ Phase0/src/worksheet_maker/WorksheetGenerator.java
@@ -1,5 +1,7 @@
 package worksheet_maker;
 
+import equation_builders.*;
+
 /**
  * Generates a worksheet through the WorksheetInput interface.
  *
